+++ src/git_secret_protector/services/key_rotator.py
@@ -1,12 +1,16 @@
 import logging
-from git_secret_protector.aes_encryption_handler import AesEncryptionHandler
-from git_secret_protector.git_attributes_parser import GitAttributesParser
-from git_secret_protector.aes_key_manager import AesKeyManager
+
+import injector
+
+from git_secret_protector.core.git_attributes_parser import GitAttributesParser
+from git_secret_protector.crypto.aes_encryption_handler import AesEncryptionHandler
+from git_secret_protector.crypto.aes_key_manager import AesKeyManager
 
 logger = logging.getLogger(__name__)
 
 
 class KeyRotator:
+    @injector.inject
     def __init__(self, key_manager: AesKeyManager, git_attributes_parser: GitAttributesParser):
         self.aes_key_manager = key_manager
         self.git_attributes_parser = git_attributes_parser